fix infinite loading when component mounting
