+++ http/console/QuestionSetList.jsx
@@ -2,22 +2,10 @@ import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import QuestionSetSummary from './QuestionSetSummary.jsx';
 import { questionSetPropTypes } from './QuestionSet.jsx';
-import Spinner from '../shared/Spinner.jsx';
 import { getQuestionSets } from './thunks';
 
 class QuestionSetListDisplay extends Component {
-	componentDidMount() {
-		this.props.getQuestionSets();
-	}
 	render() {
-		const requestCategories = Object.keys(this.props.requests);
-		for (let i = 0; i < requestCategories.length; i++) {
-			const requestCategory = requestCategories[i];
-			if (this.props.requests[requestCategory].length > 0 && requestCategory !== 'create') {
-				return <Spinner />;
-			}
-		}
-
 		const summaries = [[]];
 		let lastDelay = 200;
 		this.props.questionSets.forEach((questionSet, index) => {
@@ -51,23 +39,14 @@ class QuestionSetListDisplay extends Component {
 }
 
 QuestionSetListDisplay.propTypes = {
-	getQuestionSets: PropTypes.func.isRequired,
 	questionSets: PropTypes.arrayOf(
 		PropTypes.shape(questionSetPropTypes)).isRequired,
-	requests: PropTypes.objectOf(PropTypes.any).isRequired,
 };
 
 const mapStateToProps = (state) => ({
 	questionSets: state.questionSets,
-	requests: state.requests,
 });
 
-const mapDispatchToProps = (dispatch) => ({
-	getQuestionSets: () => {
-		dispatch(getQuestionSets());
-	},
-});
-
-const QuestionSetList = connect(mapStateToProps, mapDispatchToProps)(QuestionSetListDisplay);
+const QuestionSetList = connect(mapStateToProps)(QuestionSetListDisplay);
 
 export default QuestionSetList;
